@@ -40,8 +40,8 @@
     "@angular/platform-browser": "^18.2.5",
     "@angular/platform-browser-dynamic": "^18.2.5",
     "@angular/router": "^18.2.5",
-    "@ngnestpostgres/fe-shared": "^0.0.4",
-    "rxjs": "~7.4.0",
+    "@ngnestpostgres/fe-shared": "^0.0.32",
+    "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.10"
   },
@@ -54,7 +54,7 @@
     "@angular-eslint/template-parser": "^18.0.1",
     "@angular/cli": "^18.2.5",
     "@angular/compiler-cli": "^18.2.5",
-    "@types/jasmine": "~3.10.0",
+    "@types/jasmine": "~5.1.0",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "^7.12.0",
     "@typescript-eslint/parser": "^7.12.0",
@@ -64,13 +64,13 @@
     "eslint-plugin-simple-import-sort": "^12.1.0",
     "eslint-plugin-unused-imports": "^3.2.0",
     "husky": "^9.0.11",
-    "jasmine-core": "~3.10.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "ng-packagr": "^18.0.0",
-    "typescript": "~5.4.5"
+    "jasmine-core": "~5.2.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "ng-packagr": "^18.2.1",
+    "typescript": "~5.5.2"
   }
 }
