@@ -1,5 +1,8 @@
-import { Component, Inject, OnInit } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import {
+  ChangeDetectionStrategy, Component, inject, OnInit
+} from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
 import { AuthService } from '@auth/services/auth.service';
 import { UserLogin } from '@ngnestpostgres/fe-shared';
 // import { UserRole } from '@ngnestpostgres/fe-shared';
@@ -7,21 +10,25 @@ import { UserLogin } from '@ngnestpostgres/fe-shared';
 @Component({
   selector: 'anp-login-dialog',
   templateUrl: './login-dialog.component.html',
-  styleUrls: ['./login-dialog.component.scss']
+  styleUrls: ['./login-dialog.component.scss'],
+  standalone: true,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [
+    MatButtonModule,
+    MatDialogModule,
+  ]
 })
 export class LoginDialogComponent implements OnInit {
-  constructor(
-    private authService: AuthService,
-    private dialogRef: MatDialogRef<LoginDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) private data: unknown,
-  ) {}
+  private authService = inject(AuthService);
+  private data: unknown = inject(MAT_DIALOG_DATA);
+  private dialogRef = inject(MatDialogRef<LoginDialogComponent>);
 
   ngOnInit(): void {
     console.log('UserRole.Admin');
   }
 
   public login(creds: UserLogin) {
-    this.authService.login(creds).subscribe((token) => {
+    this.authService.login(creds).subscribe((token: string) => {
       this.closeDialog(token);
     });
   }
