@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatToolbarModule } from '@angular/material/toolbar';
 import { BrowserModule } from '@angular/platform-browser';
 import { CoreModule } from '@core/core.module';
-import { SharedModule } from '@shared/shared.module';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { AuthModule } from './auth/auth.module';
 import { HomeComponent } from './home/home/home.component';
 import { NotFoundComponent } from './home/not-found/not-found.component';
 import { SearchBoxComponent } from './home/search-box/search-box.component';
@@ -20,10 +22,12 @@ import { TopMenuComponent } from './home/top-menu/top-menu.component';
   imports: [
     BrowserModule,
     CoreModule,
-    SharedModule,
-    AuthModule,
     AppRoutingModule,
     HomeComponent,
+    MatButtonModule,
+    MatIconModule,
+    MatSidenavModule,
+    MatToolbarModule,
     NotFoundComponent,
     SearchBoxComponent,
     SidenavTreeComponent,
