@@ -14,8 +14,11 @@ describe('LoginDialogComponent', () => {
     const spyAuthService = jasmine.createSpyObj('AuthService', ['login']);
 
     await TestBed.configureTestingModule({
-      imports: [MatDialogModule, NoopAnimationsModule],
-      declarations: [LoginDialogComponent],
+      imports: [
+        LoginDialogComponent,
+        MatDialogModule,
+        NoopAnimationsModule
+      ],
       providers: [
         { provide: AuthService, useValue: spyAuthService },
         { provide: MatDialogRef, useValue: {} },
