@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { LoginDialogComponent } from '@auth/login-dialog/login-dialog.component';
 import { Observable } from 'rxjs';
@@ -7,14 +7,14 @@ import { Observable } from 'rxjs';
   providedIn: 'root'
 })
 export class DialogsService {
+  private readonly dialog = inject(MatDialog);
+
   private readonly defaultDialogConfig: MatDialogConfig = {
     panelClass: 'ta-dialog',
     disableClose: true,
-    minWidth: '320px',
-    width: '320px',
+    minWidth: '640px',
+    width: '640px',
   };
-
-  constructor(private dialog: MatDialog) { }
 
   public login(inputData?: unknown): Observable<string> {
     return this.dialog.open(LoginDialogComponent, {
