@@ -1,4 +1,5 @@
 import {
+  ChangeDetectionStrategy,
   Component, EventEmitter, Input, Output
 } from '@angular/core';
 
@@ -7,7 +8,9 @@ import { CoffeesFilter } from '../interfaces/coffees-filter';
 @Component({
   selector: 'anp-table-filters',
   templateUrl: './table-filters.component.html',
-  styleUrls: ['./table-filters.component.scss']
+  styleUrls: ['./table-filters.component.scss'],
+  standalone: true,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TableFiltersComponent {
   @Input() widthPx: number | undefined;
